crawl_app/app/requirements.txt
bs4
requests
kafka-python
html5lib
pymongo